Unite `test_to_encode_object` to `test_rlp_bytes`
`.to_encode_object()` is useful for Python to compose complex type to
make a rlp-encoded value, but it should be a only an intermediate way.
Its specific type shouldn’t be important. `bytes` or `string` or `int`,
they all become just a bytestream in RLP. So it should not be a subject
to testing. Its RLP encoded structure is a subject to testing.